third_party/blink/web_tests/external/wpt/editing/run/forwarddelete_4001-5000-expected.txt

This is a testharness.js-based test.
Found 14 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["forwarddelete",""]] "<p style=color:blue>foo[]</p>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"color:rgb(0, 0, 255)\\">foo</span>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","true"],["forwarddelete",""]] "<p style=color:blue>foo[]</p>bar" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["forwarddelete",""]] "<p style=color:blue>foo[]</p>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"#0000ff\\">foo</font>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "foo[]<p style=color:brown>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#a52a2a\\">bar</font>" but got "foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "foo[]<p style=color:brown>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#a52a2a\\">bar</font>" but got "foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span>"
[FAIL] [["stylewithcss","true"],["forwarddelete",""]] "<div style=color:blue><p style=color:green>foo[]</div>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p><span style=\\"color:rgb(0, 128, 0)\\">foo</span>bar</p></div>" but got "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foobar</p></div>"
[FAIL] [["stylewithcss","false"],["forwarddelete",""]] "<div style=color:blue><p style=color:green>foo[]</div>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p><font color=\\"#008000\\">foo</font>bar</p></div>" but got "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foobar</p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<div style=color:blue><p style=color:green>foo[]</div><p style=color:brown>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foo<font color=\\"#a52a2a\\">bar</font></p></div>" but got "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<div style=color:blue><p style=color:green>foo[]</div><p style=color:brown>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foo<font color=\\"#a52a2a\\">bar</font></p></div>" but got "<div style=\\"color:rgb(0, 0, 255)\\"><p style=\\"color:rgb(0, 128, 0)\\">foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p style=color:blue>foo[]<div style=color:brown><p style=color:green>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"color:rgb(0, 0, 255)\\">foo<font color=\\"#008000\\">bar</font></p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foo<span style=\\"color:rgb(0, 128, 0)\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p style=color:blue>foo[]<div style=color:brown><p style=color:green>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"color:rgb(0, 0, 255)\\">foo<font color=\\"#008000\\">bar</font></p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foo<span style=\\"color:rgb(0, 128, 0)\\">bar</span></p>"
[FAIL] [["forwarddelete",""]] "<p>foo</p><p>{bar</p>}<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><br></p><p>baz</p>" but got "<p>foo</p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["forwarddelete",""]] "<p>foo</p>{<p>bar}</p><p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><br><p>baz</p>" but got "<p>foo</p><p><br></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["forwarddelete",""]] "<p>foo</p>{<p>bar}</p><p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><br><p>baz</p>" but got "<p>foo</p><p><br></p><p>baz</p>"
Harness: the test ran to completion.
